requirements-dev.txt
# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
pytest-homeassistant-custom-component>=0.13.0

# Linting and Formatting
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pylint>=2.17.0

# Type Checking
mypy>=1.0.0
types-requests>=2.28.0

# Home Assistant
homeassistant>=2024.1.0

# Integration Dependencies
google-cloud-speech>=2.0.0
google-cloud-texttospeech>=2.0.0
google-generativeai>=0.3.0
voluptuous>=0.13.0

# Optional Dependencies (for full testing)
boto3>=1.26.0
azure-cognitiveservices-speech>=1.31.0
vosk>=0.3.45 